Preserving ForAll quantifiers in output CoreFn AST (WIP/maybe broken)
