+++ src/Language/PureScript/CodeGen.hs
@@ -5,4 +5,4 @@
 --
 module Language.PureScript.CodeGen (module C) where
 
-import Language.PureScript.CodeGen.JS as C
+import Language.PureScript.CodeGen.UPLC as C
