--- src/Language/PureScript/CoreFn/Typed/Module.hs
@@ -1,25 +1,0 @@
-module Language.PureScript.CoreFn.Typed.Module where
-
-import Prelude
-
-import Data.Map.Strict (Map)
-
-import Language.PureScript.AST.SourcePos (SourceSpan)
-import Language.PureScript.Comments (Comment)
-import Language.PureScript.CoreFn.Typed.Expr (Bind)
-import Language.PureScript.Names (Ident, ModuleName)
-
--- |
--- The CoreFn module representation
---
-data Module a = Module
-  { moduleSourceSpan :: SourceSpan
-  , moduleComments :: [Comment]
-  , moduleName :: ModuleName
-  , modulePath :: FilePath
-  , moduleImports :: [(a, ModuleName)]
-  , moduleExports :: [Ident]
-  , moduleReExports :: Map ModuleName [Ident]
-  , moduleForeign :: [Ident]
-  , moduleDecls :: [Bind a]
-  } deriving (Functor, Show)
